feat: setup for feature 02
