+++ src/app.js
@@ -1,15 +1,8 @@
 import React from 'react';
 
-class Header extends React.Component {
-  render() {
-    return (
-      <header>
-        <h1>City Explorer</h1>
-        <p>Enter a location below to learn about the weather, events, restaurants, movies filmed there, and more!</p>
-      </header>
-    )
-  }
-}
+import Header from './header.js';
+import Map from './map.js';
+import SearchResults from './search-results.js';
 
 class Main extends React.Component {
   render() {
@@ -21,13 +14,7 @@ class Main extends React.Component {
         <Map />
         <h2 class="query-placeholder"></h2>
         <ErrorMessage />
-        <div class="column-container hide">
-          <Results />
-          <Results />
-          <Results />
-          <Results />
-          <Results />
-        </div>
+        <SearchResults />
       </main>
     )
   }
@@ -67,14 +54,6 @@ class Search extends React.Component {
   }
 }
 
-class Map extends React.Component {
-  render() {
-    return (
-      <img id="map" class="hide" src="https://maps.googleapis.com/maps/api/staticmap?center=47.606210%2c%20-122.332071&zoom=13&size=600x300&maptype=roadmap%20%20&key=AIzaSyBfOxvSAEhF0bINfqhSTthhNKEBb8eHfHc" alt="Map of search query" />
-    )
-  }
-}
-
 class ErrorMessage extends React.Component {
   render() {
     return (
@@ -83,24 +62,13 @@ class ErrorMessage extends React.Component {
   }
 }
 
-class Results extends React.Component {
-  render() {
-    return (
-      <section>
-        <h3>Results from the Lorem Ipsum</h3>
-        <ul class="Lorem Ipsum"></ul>
-      </section>
-    )
-  }
-}
-
 class App extends React.Component {
   render() {
     return (
-      <React.Fragment>
+      <>
         <Header />
         <Main />
-      </React.Fragment>
+      </>
     );
   }
 }
