+++ src/result.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+class Result extends React.Component {
+  render() {
+    return (
+      <section>
+        <h3>Results from the Lorem Ipsum</h3>
+        <ul class="Lorem Ipsum"></ul>
+      </section>
+    )
+  }
+}
+
+export default Result;
